update `DataTable` schema and helpers for configuring csv export
- update specs and helpers for `DataTable` component
- add data export test case
- update data table demo page descriptor
- minor edits

+++ test/recipe/Baked.Test.Recipe.Service/Theme/TableRow.cs
@@ -1,3 +1,3 @@
 ﻿namespace Baked.Test.Theme;
 
-public record TableRow(string Label, string Column1, int Column2, int Column3);
+public record TableRow(string Label, string Column1, int Column2, double Column3);
